Move Helper task to a command
