+++ src/Plugin.php
@@ -20,6 +20,7 @@ use Bake\Command\CellCommand;
 use Bake\Command\CommandCommand;
 use Bake\Command\ComponentCommand;
 use Bake\Command\FormCommand;
+use Bake\Command\HelperCommand;
 use Bake\Command\TestCommand;
 use Bake\Shell\BakeShell;
 use Cake\Console\CommandCollection;
@@ -73,6 +74,7 @@ class Plugin extends BasePlugin
         $commands->add('bake command', CommandCommand::class);
         $commands->add('bake component', ComponentCommand::class);
         $commands->add('bake form', FormCommand::class);
+        $commands->add('bake helper', HelperCommand::class);
         $commands->add('bake test', TestCommand::class);
 
         // Add shell for incomplete tasks and backwards compatibility discover.
